added DatabaseException and ResourceNotFoundException as well as ResourceExceptionHandler to handle custom exceptions

+++ src/main/java/com/eduardoscheffer/oopjavawebservice/services/UserService.java
@@ -1,8 +1,12 @@
 package com.eduardoscheffer.oopjavawebservice.services;
 
 import com.eduardoscheffer.oopjavawebservice.entities.User;
+import com.eduardoscheffer.oopjavawebservice.exceptions.DatabaseException;
+import com.eduardoscheffer.oopjavawebservice.exceptions.ResourceNotFoundException;
 import com.eduardoscheffer.oopjavawebservice.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,7 +25,7 @@ public class UserService {
 
     public User findUserById(Long id) {
         Optional<User> obj = repository.findById(id);
-        return obj.get();
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
     public User insertUser(User user) {
@@ -29,7 +33,12 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
-        repository.deleteById(id);
+        try {
+            if (!repository.existsById(id)) throw new ResourceNotFoundException(id); // se nao existir o User na DB lanca uma excecao peronalizada
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     @Transactional
